+++ tests/test_core.py
@@ -388,6 +388,7 @@ def test_list_subtitles_movie(movies, mock_providers):
 
     # test providers
     assert not provider_manager['addic7ed'].plugin.list_subtitles.called
+    assert not provider_manager['cinemast'].plugin.list_subtitles.called
     assert provider_manager['legendastv'].plugin.list_subtitles.called
     assert provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert provider_manager['podnapisi'].plugin.list_subtitles.called
@@ -409,6 +410,7 @@ def test_list_subtitles_episode(episodes, mock_providers):
 
     # test providers
     assert provider_manager['addic7ed'].plugin.list_subtitles.called
+    assert not provider_manager['cinemast'].plugin.list_subtitles.called
     assert provider_manager['legendastv'].plugin.list_subtitles.called
     assert provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert provider_manager['podnapisi'].plugin.list_subtitles.called
@@ -430,6 +432,7 @@ def test_list_subtitles_providers(episodes, mock_providers):
 
     # test providers
     assert provider_manager['addic7ed'].plugin.list_subtitles.called
+    assert not provider_manager['cinemast'].plugin.list_subtitles.called
     assert not provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert not provider_manager['podnapisi'].plugin.list_subtitles.called
     assert not provider_manager['thesubdb'].plugin.list_subtitles.called
@@ -448,6 +451,7 @@ def test_list_subtitles_episode_no_hash(episodes, mock_providers):
 
     # test providers
     assert provider_manager['addic7ed'].plugin.list_subtitles.called
+    assert not provider_manager['cinemast'].plugin.list_subtitles.called
     assert provider_manager['legendastv'].plugin.list_subtitles.called
     assert provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert provider_manager['podnapisi'].plugin.list_subtitles.called
@@ -469,6 +473,7 @@ def test_list_subtitles_no_language(episodes, mock_providers):
 
     # test providers
     assert not provider_manager['addic7ed'].plugin.list_subtitles.called
+    assert not provider_manager['cinemast'].plugin.list_subtitles.called
     assert not provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert not provider_manager['podnapisi'].plugin.list_subtitles.called
     assert not provider_manager['thesubdb'].plugin.list_subtitles.called
@@ -490,6 +495,7 @@ def test_download_subtitles(mock_providers):
 
     # test providers
     assert provider_manager['addic7ed'].plugin.download_subtitle.called
+    assert not provider_manager['cinemast'].plugin.download_subtitle.called
     assert not provider_manager['legendastv'].plugin.download_subtitle.called
     assert not provider_manager['opensubtitles'].plugin.download_subtitle.called
     assert not provider_manager['podnapisi'].plugin.download_subtitle.called