+++ tests/test_extensions.py
@@ -9,21 +9,22 @@ def test_registrable_extension_manager_all_extensions():
         'de7cidda = subliminal.providers.addic7ed:Addic7edProvider'
     ])
     extensions = sorted(e.name for e in manager)
-    assert len(extensions) == 9
-    assert extensions == ['addic7ed', 'argenteam', 'de7cidda', 'legendastv', 'opensubtitles', 'podnapisi', 'shooter',
-                          'thesubdb', 'tvsubtitles']
+    assert len(extensions) == 10
+    assert extensions == ['addic7ed', 'argenteam', 'cinemast', 'de7cidda', 'legendastv', 'opensubtitles', 'podnapisi',
+                          'shooter', 'thesubdb', 'tvsubtitles']
 
 
 def test_registrable_extension_manager_internal_extension():
     manager = RegistrableExtensionManager('subliminal.test_providers', [
         'addic7ed = subliminal.providers.addic7ed:Addic7edProvider',
+        'cinemast = subliminal.providers.cinemast:CinemastProvider',
         'opensubtitles = subliminal.providers.opensubtitles:OpenSubtitlesProvider',
         'podnapisi = subliminal.providers.podnapisi:PodnapisiProvider',
         'thesubdb = subliminal.providers.thesubdb:TheSubDBProvider',
         'tvsubtitles = subliminal.providers.tvsubtitles:TVsubtitlesProvider'
     ])
-    assert len(list(manager)) == 5
-    assert len(manager.internal_extensions) == 5
+    assert len(list(manager)) == 6
+    assert len(manager.internal_extensions) == 6
 
 
 def test_registrable_extension_manager_register():
@@ -39,14 +40,15 @@ def test_registrable_extension_manager_register():
 
 def test_registrable_extension_manager_unregister():
     manager = RegistrableExtensionManager('subliminal.test_providers', [
+        'cinemast = subliminal.providers.cinemast:CinemastProvider',
         'thesubdb = subliminal.providers.thesubdb:TheSubDBProvider',
         'tvsubtitles = subliminal.providers.tvsubtitles:TVsubtitlesProvider'
     ])
-    assert len(list(manager)) == 2
+    assert len(list(manager)) == 3
     manager.register('de7cidda = subliminal.providers.addic7ed:Addic7edProvider')
     manager.unregister('de7cidda = subliminal.providers.addic7ed:Addic7edProvider')
-    assert len(list(manager)) == 2
-    assert set(manager.names()) == {'thesubdb', 'tvsubtitles'}
+    assert len(list(manager)) == 3
+    assert set(manager.names()) == {'cinemast', 'thesubdb', 'tvsubtitles'}
 
 
 def test_provider_manager():